멘토 등록기능 및 멘토 썸네일 클릭 시 상세페이지 이동 구현

+++ app/src/main/java/preview/android/model/MentorPost.kt
@@ -11,5 +11,5 @@ data class MentorPost(
     val comment : String = "",
     val subTitle : String = "",
     val categortId : Int = 1,
-    val kakaoId : Int = 0,
+    val kakaoId : Long = 0,
 ) : Serializable
